+++ src/main/java/top/stillmisty/xiantao/infrastructure/mapper/UserMapper.java
@@ -11,8 +11,8 @@ public interface UserMapper extends BaseMapper<User> {
 
   @Update(
       "UPDATE xt_user SET spirit_stones = spirit_stones - #{cost} WHERE id = #{userId} AND spirit_stones >= #{cost}")
-  int deductSpiritStonesIfEnough(@Param("userId") Long userId, @Param("cost") int cost);
+  int deductSpiritStonesIfEnough(@Param("userId") Long userId, @Param("cost") long cost);
 
   @Update("UPDATE xt_user SET spirit_stones = spirit_stones + #{amount} WHERE id = #{userId}")
-  int addSpiritStonesAtomically(@Param("userId") Long userId, @Param("amount") int amount);
+  int addSpiritStonesAtomically(@Param("userId") Long userId, @Param("amount") long amount);
 }
